Rework crystal stabilization
Add a platform/peripheral-specific constant so we don't hard-code the PUC
MCLK rate.  Add a new configuration value so we can select how long to wait
for crystal stabilization.  Refine the interface so that if LFXT1
stabilization is requested for zero loops, the system just disables the
crystal without even testing it.