Eliminate structure overloading in timer HAL state
Putting the real state structure inside a private structure prevents us from
having a global symbol that we could use as a const initializer for a HAL
reference.  Result is one additional load at ISR time, which isn't great,
but it will let us eliminate all the global variables that are pointers to
HAL states, which will eliminate a bunch of loads through those pointers.