Eliminate serial configure callbacks
The implementation failed to account for preserving and restoring existing
chain members.  It's unlikely anybody will dynamically change interrupt
chains, but if they do they'll have to manage it themselves; adding a new
element at the head of the chain is sufficiently trivial that there's no
need for a helper function.