Include platform-specific header after all other headers
We want to include the platform-specific header as a check to make sure
everything's right for the platform, but in fact it is normally included
automatically as a result of <bsp430/core.h>.  Including it too early breaks
the logic used for default resource management.  In particular, exp430g2
will fail to correctly detect the platform default timer and will not enable
the uptime feature when it should.