close #5: fix uptime resource on EXP430G2
Add a new coordination standard BSP430_feature_USE_PLATFORM_RESOURCE to
influence the actions when configBSP430_feature_USE_DEFAULT_RESOURCE is true
but the platform prefers an alternative to the generic resource normally
defined in the platform-independent headers.

Doing this also cleans up an inconsistency managing the uptime resource,
where the platform-specific platform.h improperly defined BSP430_UPTIME
instead of deferring to the standard logic in the generic platform.h.