cli: change interface to next token function
The original interface would not permit extracting tokens that have quotes
in a way that eliminated the quotes from the returned value.  It also forced
the caller to do bookkeeping on the pointer and length after the function
returned, if the token was consumed.  Most times the token will be consumed,
so it's better to make the caller preserve the old pointer/count in the case
where it isn't.