cli: prevent null command link pointers, and put command set into command link
Without recording the valid command set in the command link node, and
passing that node to handlers and diagnostics, it's not possible to
determine what alternatives might be available at a given level.  By putting
the set in the command link we also eliminate the need to pass it explicitly
to the diagnostic function.