Fix vBSP430serialWakeupTransmit_ni
The implementation had some residual confusion that mistakenly tried to set
the interrupt enable only if the IFG was already set.  IE should be set
unconditionally; if IFG is also set, it'll fire right away, otherwise it'll
fire when IFG does get set.

This does mean the infrastructure needs to be careful about keeping IFG set
when IE is clear, if in fact the peripheral is ready to transmit, but the
interrupt handlers already take care to do that.