expandtemplate: provide C preprocessor constants for core resource instances
We want to be able to do something like:

#define BSP430_UPTIME_TIMER BSP430_PERIPH_TB0

and then automatically set configBSP430_HAL_TB0 and related flags.  That
BSP430_PERIPH_TB0 is a compile-time constant, but cannot be used in C
preprocessor conditionals, makes this messy.  Provide an alternative handle
that is safe for use in preprocessor conditionals.