close #24: support XT2 configuration
Pretty much a clone-and-own of the LFXT1 configuration infrastructure.
Validated on the EXP430F5529.  CC430 boards have external 26MHz crystals,
and they show up enabled and not faulted, but basic tests failed to verify
that the crystal is working, either because it's not, it can't be routed to
an external pin, or it's too fast for my meter.