trxeb/timer_test: add example
This code was used to determine the behavior of the #SCS bit in 5xx Timer_B
implementations when the timer clock is asynchronous to MCLK.  The take-away
is that SCS means a capture will occur on the next falling edge of the timer
clock, even if the input signal has returned to its normal state beforehahd.
Where the timer clock is asynchronous to and slower than MCLK this means
toggling CCIS0 when CCIS1 is set will not necessarily load the capture
register with the current counter value, as is suggested in the user's
guide.

See http://e2e.ti.com/support/microcontrollers/msp43016-bit_ultra-low_power_mcus/f/166/t/236172.aspx