close #31: provide reliable read of asynchronous timer counters
Reading an MSP430 timer counter when the clock is asynchronous to MCLK can
return invalid results.  Reserve CC2 to safely read the counter of a timer
by capturing a valid value regardless of whether the clock is synchronous or
asynchronous to MCLK, and whether it is faster or slower than MCLK.

As best as possible reconfiguration of the capture/compare register will cause the application to hang at point-of-use,
allowing diagnosis during basic development.

On an MSP430F5438A, it used to take 60 cycles to read the 32-bit uptime
counter.  With this change it now takes 68 cycles.  Small cost compared to
trying to diagnose an intermittent bug due to a corrupt counter read in the
alarm infrastructure.