uptime: do not interpret BSP430_UPTIME_DELAY_CCIDX for feature-enable purposes
Using a non-zero value of the CCIDX macro as the cue that the delay feature
is enabled unnecessarily prevents use of CC0 for this where RTOS support is
not needed, and leads to confusion in bsp430_config.h files.  Follow
standard practice of using a configBSP430_FOO flag.

Also refine the description of why configBSP430_CORE_LPM_EXIT_CLEAR_GIE
should be used.