ucs: default to configBSP430_CORE_DISABLE_FLL (closes #18)
Issue #18 was determined to be related to short wake times processing
multi-character 115200baud UART receptions.  Stack traces showed the
application to be executing on the stack in this situation.  The anomalous
behavior disappeared when configBSP430_CORE_DISABLE_FLL was set, so the
finger of blame is being pointed at erratum UCS7.