close #51: support on-board LCDs
u8glib turns out to be usable if the provided display devices are ignored
so that AVR-related noise isn't pulled in.  It doesn't seem worthwhile to
try to create a communications module that can be shared and integrated into
u8glib since the SPI configuration varies between devices.

This is really just a proof-of-concept.  For any real-world LCD applications
that might want to do partial updates (such as a scrolling oscilloscope
trace) the u8glib interface is possibly too page/screen-oriented.