rfem refactoring
Hindsight revealed that using EMK pin functions as the defining
characteristic for the RFEM interface was a mistake: too many newer EMKs
don't follow the CC2500/CC110x naming conventions.  Rework this so the RFEM
module associates port and timer peripheral information with specific RF
header pins on a per-platform/per-mcu basis using tables to generate the
bindings automatically.

Along with this there are maps from the RFEM pins to the functional signals
for a set of particular radios, currently Chipcon ones.  The interface here
is only the core hardware presentation required to access the radio signals;
BSP430 is not intended to support radio-specific operations such as
transmitting packets or even reading and writing configuration registers.
There are crude examples for each radio that interact with it enough to
extract the radio identifier; more complex operations are the responsibility
of a radio driver that is focused on the radio itself rather than the host
platform.

This makes it more difficult to share code between CC2500/CC110x
implementations, while enabling CC112x and CC2520.  All in all a fair
trade.  (There is stub information for the PAN1323ETU but it has not been
integrated or tested.)