timer: add pulse width measurement infrastructure
Functions and data structures to measure the width of a pulse received on a
timer capture signal.  Somewhat similar to the alarm functionality in that
the interrupt callbacks are registered only when pulse capture is enabled.
Overflow is detected both within the timer peripheral and if the application
does not extract the pulse data before another transition occurs.
Application is notified on overflows and optionally at the start and end of
the pulse.