resource: refine handling of waiter (documentation and code)
Clearly document that waiter position in queue is only affected by the
initial resource claim.  Also invoke new head callback when the wait list
changes, in case the list changed due to a cancel from within a callback (so
the release is not lost) or from outside (when the new head may recursively
claim the resource).