timer: rework overflow function
In almost every case the overflow value will be combined with a
peripheral counter value to produce an extended counter.  Without
knowing what that peripheral counter value is we couldn't tell whether
it was probably obtained before or after the unhandled overflow event.
Assume that if the counter value is small it was read after the
overflow.