timer: optimize counter overflow handling
The recent facility to condition adjustment for unhandled overflow count by
checking value of a recent counter read eliminates the need to check for a
match in the OFIFG either side of reading the counter low word.  Also make
the overflow adjustment inline so it can be optimized at point of use and
won't cost a function call.