console: eliminate interrupt disabling for console operations
The impact of this on timely interrupt processing was prohibitive.
Assume that applications take care not to interleave output between
different tasks; if this is a possibility the resource framework
supports a solution.  Assume that applications do not need to produce
console output from interrupts, and that temporary output used during
debugging can tolerate the potential for interleaved output.