close #64: remove unnecessary @def annotations
For the record: the only place @def is required is when there are multiple
related macros that have alternative definitions and for convenience the
definitions are grouped within preprocessor conditionals.  Then it's
necessary to associate the documentation with the name because the next
macro to be defined might not be the one that was just documented.

Turns out BSP430 doesn't have any of those cases.