eui64: support application-specific, platform-specific, or generic
configBSP430_EUI64_USE_PLATFORM is intended to specify use of the
platform-provided EUI64 routine (as opposed to one provided by the
application), and did not adequately distinguish the case where the platform
has a custom implementation.  Now that there are other functions in the
utility/eui64 module that distinction is important.