pinmaps: support BoosterPack mapping
Adds maps for standard booster packs enabling the EM Adapter BoosterPack
to be supported on the EXP430F5529LP (nominally).

This may not be entirely useful: the SENSHUB RFEM interface maps to pins
that don't support SPI on the MSP430 booster pack standard; the CCEM
booster pack maps the most GPIO lines to ports that don't support
interrupts on the EXP430F5529LP; the EXP430FR5739 already has an RFEM
socket; and the other boards with BoosterPack capability don't have
enough memory to be useful.