maintainer: add showpins utility
This utility, which does not have a user-facing interface, can be used
to display tables that relate radio or booster-pack functions through
header and MCU pins down to MSP430 peripherals, to see whether a
boosterpack/platform combination has a chance of functioning.

In theory, the CC3000BOOST will work on an EXP430F5529LP.  The CC3000EM
in the SensHub RFEM socket will not.