cc3000boost: add support
Verified to work on the EXP430F5529LP.  Compiles on the EXP430G2 but
fails to start, probably overrunning the stack, but I'm leaving the
configuration in place just in case.

Note: The CC3000BOOST is the only option for the EXP430F5529LP.  The
RFEM socket on the SENSHUB doesn't align with SPI-capable pins, and the
CCEM puts the SPI IRQ signal on P6.5 which doesn't support interrupts.
Does anybody check these designs?