bootstrap/clocks: add CS variant info to output and blink
In the CS peripheral output indicate whether this is a CS (FR57xx) or
CS_A (FR58xx) variant.

To assist in debugging issues (like setting a clock speed that
immediately generates an oscillator fault which reboots the board,
versus a clock speed for which UART output is unreliable), blink the
first LED in the idle loop.