uptime: simplify use of delay when interrupts are enabled
In many cases the context of a delay does not inherently require that it
be non-interruptible.  Avoid point-of-use interrupt disabling by
providing alternative macros that provide the uninterruptible context
this where an exit condition is checked, and by having the underlying
primitive save and restore the interrupt state so the internal need for
an uninterruptible context does not impact the external context.