timer: support reading full-precision times and correcting them
As we start to do more with NTP, having the full-precision available
only as a 32-bit low and 16-bit overflow counter is inconvenient.
Provide a way to return the 48-bit counter as a 64-bit integer.

Also provide a way to adjust that value so that a full-precision
representation of a timestamp captured in hardware as a 16-bit counter
has the correct corresponding setting for the upper 48 bits regardless
of subsequent overflow (so long as the 16-bit counter did not lap
itself).