utility/event: refine description of periodic event fields
A casual reader might think it possible to change the interval to stop
repetition after the next event.  While theoretically this might be
useful, it's best not to mutate the values of fields read in an ISR, so
disallow it.  Implement timely processing of the event and cancel it
there, or use a non-event muxalarm callback.