eui64: refine null/invalid value management
Not sure why the all-1s EUI-64 was used for NULL, as the all-0s version
makes more sense.  Codify use of the all-1s EUI-64 to denote an invalid
value as opposed to an unassigned value; this is more consistent with
the practice of zeroing a structure to create an initial state.