uptime: add utilities for identifying and adjusting special uptime values
Special identifying values for uptime might occur in a couple
situations: when using an interrupt-captured signal to transmit local
time to another node through an RF interface, an initialized value of
zero might indicate that the captured time was not written into the
packet before transmission occurred.

Analogously, an initialized value of ~0UL might indicate that sensor
samples corresponding to that time are not available.

Provide a facility to distinguish times from these identifying values,
and to adjust valid times so they do not happen to match one of these
values.