clock: refine identification of CS variant
Reconcile the existing CS_IS_FR5xxx macros with the need for CS4 macros
done at the clock level.  Also correct the refactored clock bootstrap
program to work again with the real CS module.

For lack of anywhere useful to put it: msp430-elf doesn't work for
bootstrap/clocks on the EXP430FR5739 because it doesn't have enough RAM
for newlib's printf so it hangs in _bsp430_sbrk_error.